Port revisions 25241:25244 from the ZopeX3-3.0 branch.  Only revisions 25242 and 25244 apply. --------------------------------------------------------------------------- Log message for revision 25242: Changes to access logging: - it's called "access" logging, not "hit" logging - keep the "hitlog" configuration section, but generate a   DeprecationWarning telling that the "accesslog" section is preferred
Response to checkin message from revision 25133, which introduced <hitlog>:

I wish I'd had time to review the original introduction of <hitlog>
when it went in; it would not be where it is now.  The changes to
ZConfig will not be eternalized to the ZConfig project; I'll have to
change things around before the second beta.

Hopefully I'll be able to review zope.server.logger at that point as
well.
---------------------------------------------------------------------------
Log message for revision 25244:
add an <accesslog> section to the template configuration file